HomeKit  QR-Code
Current problem:
HomeKit's QR-Code registration requires that the DeviceID and SetupHash information be announced in mDNS.
The SetupHash is a string Hash value that combines the SetupID and DeviceID.
In the current code, the SetupHash is a Hash value of DeviceID only.
For this reason, QR-Code reading does not recognize it as a HomeKit accessory.

Suggested fixes:
1. Add setup_id to homekit's option in the config file and add it when SetupHash is calculated.
2. Add API to get string for QR-Code using these values.
3. Add a HomeKit section to the web's links page.