fix: redact passwords in logs
Fixes https://github.com/AlexxIT/go2rtc/issues/238

Replaces URLs of the format `rtsp://user:password@localhost:8554` with
`rtsp://user:xxxxx@localhost:8554` in logs. This is best-effort for now
and does not handle cases where passwords appear in query strings. It
should be fairly easy to extend the `RedactPassword` function in the
future in case there are other common password pattern that are worth
handling.